change http-server-one recv buffer to be larger
a 1024 `recv` works fine for incoming GET requests but can cause 400s when working with a POST body that can vary. 

We are already approaching 1024 before the POST body is tacked on to the request so this can result in a truncation of the incoming request and could potentially chop off the POST body resulting in a 400 from the server since you wont be able to parse out that body.

`recv` needs to stay as a power of 2, moving the buffer up to 4096 should handle most cases for the scope of this assignment and more flexibility for creativity students may introduce beyond the scope.